Add files via upload
So here is its application, the inspiration for this came from a game called SIMONS
 
The rules of the game are really simple.
Firstly, the game shows the first colour in the sequence (can be anything). The user clicks on that button.
Next, the game shows the next colour, the user has to remember the sequence i.e the previous color, current color and so on and so forth.
If the user messes up the sequence, then the game ends.

Basically you have to click the color in sequence from the starting color and maintain the order.